+++ lib/src/cache/byte_storage_factory_stub.dart
@@ -1,0 +1,5 @@
+import '../io/io.dart';
+import 'byte_storage.dart';
+
+ByteStorage createByteStorage(Future<Directory> Function()? cacheFolder) =>
+    throw 'Not implemented';