+++ lib/src/io/io.dart
@@ -1,0 +1,1 @@
+export 'io_stub.dart' if (dart.library.io) 'io_io.dart';