+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 9.0.0-beta.3
+
+* Reduce data usage for contour lines.
+
+## 9.0.0-beta.2
+
+* Added support for sources of type [`raster-dem`](https://docs.mapbox.com/style-spec/reference/sources/) to produce contour lines. This feature is beta and could change.
+
+## 9.0.0-beta.1
+
+* Added web support. This feature is beta and could change.
+
+## 8.0.0
+
+* update `flutter_map` to version 7.0.2
+
 ## 7.3.1
 
 * support cancellation in raster tile provider